Setup stack guard memory region
This is a 1 MiB buffer between the highest-addressed loaded segment from
the ELF and the lowest address in the stack region.

The idea here is that due to the simplicity of our MMU, our stack flows
directly into the loaded segments, rather than running into unmapped
memory.

Unfortunately the highest loaded segment in our example programs has RW
permissions so a stack overflow would not be detected by the MMU unless
it overflowed past the data segment and into a read-only or RE segment.

This stack guard segment is marked as RAW so that, should uninitialized
memory tracking be turned on the MMU will fault on reads, and it will
always fault on writes.